### 2020-11-25 #414 @pachamaltese
 censo2017: Base de Datos de Facil Acceso del Censo 2017 de Chile (2017 Chilean Census Easy Access Database)
Persona Encargada: Pachá (@pachamaltese)
Repositorio:  https://github.com/pachamaltese/censo2017
Versión Enviada:  0.2
Archivo: TBD  2021-2-25
Versión Aceptada: TBD

---

-   Pega el archivo DESCRIPTION completo dentro del siguiente bloque de código.

```
Package: censo2017
Title: Base de Datos de Facil Acceso del Censo 2017 de Chile
    (2017 Chilean Census Easy Access Database)
Version: 0.4
        person(given = "Mauricio",
               family = "Vargas",
               email = "mvargas@dcc.uchile.cl",
        person(given = "Juan",
               family = "Correa",
               role = c("ctb")),
        person(family = "Instituto Nacional de Estadisticas (INE)",
               role = c("dtc")
        )
    )
Description: Provee un acceso conveniente a mas de 17 millones de registros
    de la base de datos del Censo 2017. Los datos fueron importados desde
    el DVD oficial del INE usando el Convertidor REDATAM creado por Pablo De
    Grande. Esta paquete esta documentado intencionalmente en castellano
    asciificado para que funcione sin problema en diferentes plataformas.
    (Provides convenient access to more than 17 million records from the
    Chilean Census 2017 database. The datasets were imported from the official
    DVD provided by the Chilean National Bureau of Statistics by using the
    REDATAM converter created by Pablo De Grande and in addition it includes the
    maps accompanying these datasets.)
License: CC0
    DBI,
    duckdb,
    httr,
    cli,
    crayon,
    rstudioapi,
    tools
    covr,
    dbplyr

```

## Alcance

- Por favor, indica qué categoría(s) aplican a este paquete. Las puedes encontrar en nuestras [políticas de inclusión de paquetes (Inglés)](https://ropensci.github.io/dev_guide/policies.html#package-categories). Por favor, tilda todas las apropiadas. Si no estás seguro, te sugerimos que comiences un pre-envío.

	- [ ] recuperación de datos (data retrieval)
	- [x] extracción de datos (data extraction)
	- [ ] munging de datos (data munging)
	- [ ] disposición o declaración de datos (data deposition)
	- [ ] automatización de flujos de trabajo (workflow automation)
	- [ ] control de versiones (version control)
	- [ ] manejo de citas y bibliometría (citation management and bibliometrics)
	- [ ] envoltorios de software científico (scientific software wrappers)
	- [ ] herramientas para trabajo de campo y reproducibilidad (field and lab reproducibility tools)
	- [ ] ligamientos con software de base de datos (database software bindings)
	- [x] datos geoespaciales (geospatial data)
	- [ ] análisis de texto (text analysis)

- Explica cómo y por qué el paquete encaja dentro de estas categorías (1 a 3 oraciones):

Este trabajo es una adaptación de `citesdb`, también parte de rOpenSci, a un contexto diferente.

Para ayudar a hacer investigación reproducible, converti la informacion desde el DVD en formato REDATAM usando el Convertidor REDATAM creado por Pablo De Grande. La modificacion a estos archivos, que incluye geometrias detalladas, consistio en unir todos los archivos shp regionales en una unica tabla por nivel (e.g en lugar de proveer R01_mapa_comunas, ..., R15_mapa_comunas combine las 15 regiones en una unica tabla mapa_comunas).

- ¿Cuál es la audiencia esperada y las aplicaciones científicas de este paquete?

Economistas, sociologos y otros profesionales que trabajan con datos demograficos.

- ¿Hay otros paquetes de R que logren el mismo objetivo? Si los hay, ¿cómo se diferencian del tuyo, o alcanzan nuestro [criterio del mejor de su categoría (documento en Inglés)](https://ropensci.github.io/dev_guide/policies.html#overlap)?

No existen otros

-   (Si aplica) ¿Tu paquete cumple con nuestras [guías de _Ética, Privacidad de Datos e Investigación de Sujetos Humanos_ (documento en Inglés)](https://devguide.ropensci.org/policies.html#ethics-data-privacy-and-human-subjects-research)?

Si, de hecho no individualiza a personas, sino que el nivel detalle es a nivel de zonas (divisiones comunales).  Existe una tabla personas, pero es para permitir obtener  el % de adultos mayores por zona, etc.

- Si ya has hecho una consulta de pre-envío, por favor pega el enlace al _issue_ correspondiente, una publicación del foro, u otra discusión. Alternativamente, etiqueta al editor (con @tag) con el que te contactaste.

https://github.com/ropensci/software-review/issues/410

## Revisiones Técnicas

Tilda los siguientes items para confirmar que los has completado:

- [x] He leído la [guía para autores (documento en Inglés)](https://devguide.ropensci.org/guide-for-authors.html) y [la guía de paquetes de rOpenSci (documento en Inglés)](https://devguide.ropensci.org/building.html).

Este paquete:

- [x] No viola los Términos de Servicio de ningún servicio con los que interactúa.
- [x] Tiene una licencia aceptada por CRAN y OSI.
- [x] Contiene un [archivo README con instrucciones para instalar la versión de desarrollo](https://ropensci.github.io/dev_guide/building.html#readme).
- [x] Incluye [documentación con ejemplos para todas las funciones, creada con roxygen2](https://ropensci.github.io/dev_guide/building.html#documentation).
- [x] Contiene una viñeta (_vignette_) con ejemplos de sus funciones esenciales y su uso.
- [x] Tiene una [suite de tests (documento en Inglés)](https://ropensci.github.io/dev_guide/building.html#testing).
- [x] Tiene [integración contínua (documento en Inglés)](https://ropensci.github.io/dev_guide/ci.html), incluyendo reporte de cobertura de tests, usando servicios como Travis CI, Coveralls y/o CodeCov.

## Opciones de Publicación

- [x] ¿Tienes intenciones de subir este paquete a CRAN?  ya esta en CRAN
- [ ] ¿Tienes intenciones de enviar este paquete a Bioconductor?

- [ ] ¿Deseas enviar un Artículo de Aplicaciones sobre tu paquete a [Methods in Ecology and Evolution (documento en Inglés)](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/)? Si es así:

<summary>Opciones para MEE</summary>

- [ ] Este paquete es novedoso y será de interés para la mayoría de lectores de la revista.
- [ ] El manuscrito que describe el paquete no tiene más de 3000 palabras y está escrito en Inglés.
- [ ] Tienes intenciones de archivar el código del paquete en un repositorio a largo plazo, que cumple los requerimientos de la revista (mira las [Políticas de Publicación de MEE (documento en Inglés)](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/journal-resources/policy-on-publishing-code.html))
- (*Alcance: Considera los [Objetivos y Alcance de MEE (documento en Inglés)](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/aims-and-scope/read-full-aims-and-scope.html) para tu manuscrito. No otorgamos garatías de que tu manuscrito esté en el ámbito de MEE.*)
- (*Aunque no es requerido, recomendamos tener un manuscrito completamente preparado y en Inglés, al momento de enviar.*)
- (*Por favor, no envíes tu paquete de forma separada a Methods in Ecology and Evolution*)

## Código de Conducta

- [x] Estoy de acuerdo en cumplir el [Código de Conducta de rOpenSci (documento en Inglés)](https://ropensci.github.io/dev_guide/policies.html#code-of-conduct) durante el proceso de revisión, y en mantener mi paquete si éste es aceptado.
