### 2020-11-25 #414 @melvidoni
### Editor checks:

---

#### Editor comments

Hola @pachamaltese, muchas gracias por pilotear las revisions en español con rOpenSci. Para empezar, estos son los resultados de `goodpractice`:

```
It is good practice to

    R/connect.R:3:NA
    R/connect.R:4:NA
    R/connect.R:5:NA
    R/connect.R:6:NA
    R/connect.R:8:NA
    ... and 178 more lines

  ✖ omit "Date" in DESCRIPTION. It is not required and it gets invalid quite often. A build date will be added to the package when you perform
    `R CMD build` on it.
  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80 characters wide. Try make your
    lines shorter than 80 characters

    R\connect.R:14:1
    R\connect.R:27:1
    R\connect.R:60:1
    R\connect.R:78:1
    R\connect.R:131:1
    ... and 7 more lines

  ✖ avoid sapply(), it is not type safe. It might return a vector, or a list, depending on the input data. Consider using vapply() instead.

    R\zzz.R:11:11
```

¿Podrías trabajar en esto y avisarme cuando este listo? Por favor, dejá una respuesta detallada con los cambios.

Por el momento, iré contactando candidatos a revisores.

---

Reviewers: TBD
Due date: TBD
