### 2020-11-28 #414 @pachamaltese
Listo!! Implementé casi todo lo que dice goodpractice, salvo lo del `sapply()` pues si implemento el cambio se rompe todo. Con respecto a los tests, las funciones que goodpractice que dice que no están cubiertas, es porque los tests existentes indirectamente llaman a esas funciones, por ejemplo https://github.com/pachamaltese/censo2017/blob/main/tests/testthat/test-02-tables.R#L12 llama a la conexion del panel de todos modos y usa `connect.R` internamente.
