### 2020-11-28 b076424
agrega funciones internas a codecov ignore
