### 2020-11-29 #414 @melvidoni
Hola @pachamaltese, muchas gracias por los cambios.

Los revisores son @anadiedrichs y @fvd. Por favor, recuerden usar la [Guía para Revisores](https://devguide.ropensci.org/reviewerguide.html) y las plantillas correspondientes para la revisión. Sabemos que las plantillas están en inglés, pero bueno... por favor, completen todo en español. Esta revisión es un piloto y aún no tenemos todo traducidor.

**Fecha Límite para la revisión: ** Dada la temporada del año, y que es 2020, sumado a una revisión piloto... la fecha límite para la revisión va a ser el 14 de Enero de 2021.
