### 2020-11-30 #414 @pachamaltese
Hola @melvidoni
¿Existirá un potencial conflicto de interés? con ambos revisores hemos colaborado en múltiples instancias (R4DS,  plumber, shiny, etc). Yo no lo veo así, creo que tendrán la confianza para enviar feedback de lo bueno y lo malo, pero siempre es bueno avisar.
