### 2020-11-30 #414 @melvidoni
Te conoce demasiada gente @pachamaltese! En fin, me parece genial que avises, por ahora también lo dejo a criterio de los revisores. Tenemos nuestra [Guía de Conflictos de Interés](https://devguide.ropensci.org/policies.html#coi). Si ellos no se consideran libres de COI, pido que me avisen, y volvemos al estadío en busca de revisores.
