### 2020-12-02 #414 @FvD
@melvidoni y @pachamaltese siguiendo la [Guía de Conflictos de Interés](https://devguide.ropensci.org/policies.html#coi) yo me siento libre de algún conflicto de interés.
