### 2020-12-02 #414 @melvidoni
> @melvidoni y @pachamaltese siguiendo la [Guía de Conflictos de Interés](https://devguide.ropensci.org/policies.html#coi) yo me siento libre de algún conflicto de interés.

Excellente. Por favor, procede con la revisión entonces.
