### 2020-12-04 #414 @stefaniebutland
Hola @FvD. I'm rOpenSci's Community Manager. If you would like to join our Slack, please send me your email address to `stefanie@ropensci.org`. I invite all package authors and reviewers.
