### 2020-12-14 #414 @melvidoni
Estimados, les recuerdo que no habrá movimiento editorial entre diciembre 19 y enero 3, como se explica en #417.
