### 2020-12-15 8f46199
CRAN requested changes
