+++ NEWS.md
@@ -1,0 +1,5 @@
+# Version 0.2
+
+- Adds `vignettes/` to `.Rbuildignore`.
+- Complies with CRAN policies regarding Suggests.
+- Vignettes are now available from gh-pages.
