### 2021-01-16 #414 @FvD
*Nota: revisión esta en proceso - aun no lo he finalizado*
*Nota 2: este machote de revisión `review-es.md` esta disponible como pull request en el `dev-guide` [#301](https://github.com/ropensci/dev_guide/issues/301)*

## Revisión del paquete

*Por favor trata de marcar tantas casillas como te es posible y elabora tus argumentos en comentarios abajo de cada uno. Tu revisión no se necesita listar a estos temas, tal como se describe en la guia para revisores (Reviewer Guide)*

**Por favor describe cualquier relación de trabajo que tienes/has tenido con los autores del paquete)**

- [x] Como revisor confirmo que no tengo [conflictos de interés](https://devguide.ropensci.org/policies.html#coi) para poder hacer la revisión de este trabajo (si no estas seguro si tienes un conflicto por favor entra en contacto con tu editor _antes_ de arrancar con la revisión.

#### Documentación

El paquete incluye todos los siguiente tipos de documentación:

- [ ] **Una declaración de necesidades** que claramente describe las necesidades que el software esta diseñado a resolver y el public meta que busca atender en el archivo README

Hay una sección de "valor agregado" pero si interpreto el requisito de rOpenSci de forma estricta, no veo una declaración de necesidades. ¿Porque trabajar en el paquete? ¿Que carencia hay en el mundo R o el mundo de datos de Censo para justificar el esfuerzo de crear este paquete? ¿Para quien lo estas escribiendo?.

Haces referencia a tu sitio web https://db-edu.pacha.dev pero esto esta en inglés. Aun si tienes que repetir muy en breve cual es la razón de hacerlo, seria bueno tenerlo aqui en español.

- [ ] **Instrucciones de instalación** tanto para la versión en desarrollo incluyendo cualquier dependencia no-estándar en el archivo README

Creo que seria bueno ser explicitos sobre la dependencias a nivel OS que este paquete tra. En particular para instalar `sf`, necesitamos satisfacer:

```
  Configuration failed because libudunits2.so was not found. Try installing:
    * deb: libudunits2-dev (Debian, Ubuntu, ...)
    * rpm: udunits2-devel (Fedora, EPEL, ...)
    * brew: udunits (OSX)
```
y necesitamos gdal que puede ser [un poco engorroso](https://stackoverflow.com/questions/15248815/rgdal-package-installation), especialmente para alguien que no lo ha hecho antes. Seria bueno tener algo de comentario sobre como satisfacer esta dependencia a nivel OS.

```
configure: error: gdal-config not found or not executable.
ERROR: configuration failed for package ‘sf’
```

- [ ] **Vignette(s)** demostrando la funcionalidad principal que ademas corren localmente

- [ ] **Documentación de las funciones** en la ayuda de R para todas la funciones exportadas
- [ ] **Ejemplos** para todas las funciones exportadas que corren localmente
- [ ] **Directrices comunitarias** incluyendo una guia de contribución en el archivo README o el archivo CONTRIBUTING y un archivo DESCRIPTION que incluye `URL`, `BugReports` and `Maintainer` (todas en inglés por concenvión y para que puedan ser autogeneradas con `Authors@R`).

#### Funcionalidad

- [ ] **Instalación:** La instalación se completa con éxito tal como fue documentado.
- [ ] **Funcionalidad:** Todo afirmación de funcionalidad del software se confirma como existente.
- [ ] **Desempeño:** Toda afirmación de desempeño del software se confirma como alcanzada.
- [ ] **Pruebas automáticas:** Hay pruebas unitarias que cubren las funciones esenciales dentro del paquete con un rango razonable de entradas y condiciones. Todas las pruebas corren en la maquina local.
- [ ] **Directrices de empaque**: El paquete cumple con las directrices de empaque de rOpenSci.

Estimación de horas dedicadas a la revisión:

- [ ] Si la o las persona(s) autora(s) lo considera(n) apropiado, yo estoy de acuerdo con que me reconozcan como revisor del paquete (el rol "rev') en la el archivo DESCRIPTION del paquete.

---
