### 2021-01-23 2be8b63
cambios pedidos por CRAN para R 4.0
