+++ NAMESPACE
@@ -14,4 +14,3 @@ importFrom(httr,progress)
 importFrom(httr,stop_for_status)
 importFrom(httr,write_disk)
 importFrom(purrr,keep)
-importFrom(rappdirs,user_data_dir)
