+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: censo2017
 Title: Base de Datos de Facil Acceso del Censo 2017 de Chile
     (2017 Chilean Census Easy Access Database)
-Version: 0.2
+Version: 0.3
 Authors@R: c( 
         person(given = "Mauricio",
                family = "Vargas",
@@ -36,7 +36,6 @@ RoxygenNote: 7.1.1
 Imports:
     DBI,
     RSQLite,
-    rappdirs,
     httr,
     R.utils,
     tibble,
@@ -44,7 +43,9 @@ Imports:
     sf,
     cli,
     crayon,
-    rstudioapi
+    rstudioapi,
+    rappdirs,
+    tools
 Suggests:
     testthat,
     covr,
@@ -52,4 +53,4 @@ Suggests:
     dplyr,
     dbplyr
 Depends: 
-    R (>= 3.5)
+    R (>= 4.0)
