### 2021-01-23 23c8cf6
elimina caracteres especiales en msje de inicio
