### 2021-01-23 #414 @pachamaltese
Estimados/as
CRAN me pidió algunos cambios que implican actualizar a R 4.0 por las políticas de escribir en el home (o en el sistema de archivos). El cambio de fondo es la ruta donde se ubica la base de datos local. Aproveché además de pasar de SQLite a duckdb, lo que aumenta fuertemente la velocidad para filtrar datos censales.
Buen finde!
