### 2021-01-28 #414 @maelle
@mpaulacaldas aqui esta la plantilla de revisión traducida por @FvD https://devdevguide.netlify.app/reviewtemplatees.html
