### 2021-02-01 bb085a4
usando duckdb 0.2.3
