+++ data-raw/01-create-local-db-duckdb.R
@@ -1,0 +1,280 @@
+# packages ----
+
+library(DBI)
+library(duckdb)
+library(RSQLite)
+
+# connections ----
+
+con <- dbConnect(SQLite(), "data-raw/censo2017.sqlite")
+tablas <- dbListTables(con)
+
+con2 <- dbConnect(duckdb(), "data-raw/censo2017.duckdb")
+
+# comunas ----
+
+dbSendQuery(con2, "DROP TABLE IF EXISTS comunas")
+
+dbSendQuery(
+  con2,
+  "CREATE TABLE comunas (
+	comuna_ref_id INTEGER NOT NULL,
+	provincia_ref_id INTEGER NULL,
+	idcomuna VARCHAR NULL,
+	redcoden VARCHAR(5) NOT NULL,
+	nom_comuna VARCHAR NULL)"
+)
+
+# hogares ----
+
+dbSendQuery(con2, "DROP TABLE IF EXISTS hogares")
+
+dbSendQuery(
+  con2,
+  "CREATE TABLE hogares (
+	hogar_ref_id INTEGER NOT NULL,
+	vivienda_ref_id INTEGER NULL,
+	nhogar INTEGER NULL,
+	tipo_hogar INTEGER NULL,
+	ncu_yern_nuer INTEGER NULL,
+	n_herm_cun INTEGER NULL,
+	nuc_herm_cun INTEGER NULL,
+	num_sueg_pad_abu INTEGER NULL,
+	nuc_pad_sueg_abu INTEGER NULL,
+	num_otros INTEGER NULL,
+	nuc_otros INTEGER NULL,
+	num_no_par INTEGER NULL,
+	nuc_no_par INTEGER NULL,
+	tot_nucleos INTEGER NULL)"
+)
+
+# mapa comunas ----
+
+dbSendQuery(con2, "DROP TABLE IF EXISTS mapa_comunas")
+
+dbSendQuery(
+  con2,
+  "CREATE TABLE mapa_comunas (
+	geometry VARCHAR NULL,
+	region VARCHAR(2) NULL,
+	provincia VARCHAR(3) NULL,
+	comuna VARCHAR(5) NOT NULL)"
+)
+
+# mapa provincias ----
+
+dbSendQuery(con2, "DROP TABLE IF EXISTS mapa_provincias")
+
+dbSendQuery(
+  con2,
+  "CREATE TABLE mapa_provincias (
+	geometry VARCHAR NULL,
+	region VARCHAR(2) NULL,
+	provincia VARCHAR(3) NOT NULL)"
+)
+
+# mapa regiones ----
+
+dbSendQuery(con2, "DROP TABLE IF EXISTS mapa_regiones")
+
+dbSendQuery(
+  con2,
+  "CREATE TABLE mapa_regiones (
+	geometry VARCHAR NULL,
+	region VARCHAR(2) NOT NULL);"
+)
+
+# mapa zonas ----
+dbSendQuery(con2, "DROP TABLE IF EXISTS mapa_zonas")
+
+dbSendQuery(
+  con2,
+  "CREATE TABLE mapa_zonas (
+	geometry VARCHAR NULL,
+	region INTEGER NULL,
+	provincia INTEGER NULL,
+	comuna INTEGER NULL,
+	geocodigo VARCHAR(11) NOT NULL)"
+)
+
+# personas ----
+
+dbSendQuery(con2, "DROP TABLE IF EXISTS personas")
+
+dbSendQuery(
+  con2,
+  "CREATE TABLE personas (
+	persona_ref_id INTEGER NULL,
+	hogar_ref_id INTEGER NULL,
+	personan INTEGER NULL,
+	p07 INTEGER NULL,
+	p08 INTEGER NULL,
+	p09 INTEGER NULL,
[…]
