+++ man/censo_borrar_base.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/download.R
+% Please edit documentation in R/remove.R
 \name{censo_borrar_base}
 \alias{censo_borrar_base}
 \title{Elimina la Base de Datos del Censo de tu Computador}
