+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: censo2017
 Title: Base de Datos de Facil Acceso del Censo 2017 de Chile
     (2017 Chilean Census Easy Access Database)
-Version: 0.3
+Version: 0.3.2
 Authors@R: c(
         person(given = "Mauricio",
                family = "Vargas",

[…]
