### 2021-02-01 b2210ee
ruta indicando version duckdb
