### 2021-02-02 b1aa5e8
actualiza nota directorio usuario, de nuevo
