### 2021-02-02 017c620
mensaje error bbdd informativo
