### 2021-02-02 5a8a14e
parche estructura de datos mapa_zonas
