### 2021-02-03 28ef46f
parche tipo cols en tabla_censo
