### 2021-02-04 da66b17
crea bbdd a partir de tsv para evitar problemas de versiones
