+++ R/connect.R
@@ -95,7 +95,7 @@ censo_bbdd <- function(dir = censo_path()) {
 #' }
 censo_tabla <- function(tabla) {
   if (any(tabla %in% grep("mapa_", censo_tables(), value = T))) {
-    df <- sf::st_read(censo_bbdd(), tabla)
+    df <- sf::st_read(censo_bbdd(), tabla, as_tibble = TRUE, quiet = TRUE)
   } else {
     df <- tibble::as_tibble(DBI::dbReadTable(censo_bbdd(), tabla))
   }
@@ -160,7 +160,7 @@ censo_estado <- function(msg = TRUE) {
 censo_tables <- function() {
   c("comunas", "hogares", "mapa_comunas", "mapa_provincias",
     "mapa_regiones", "mapa_zonas", "personas", "provincias",
-    "regiones", "viviendas", "zonas", "metadata")
+    "regiones", "viviendas", "zonas", "metadatos")
 }
 
 censo_cache <- new.env()
