+++ DESCRIPTION
@@ -37,7 +37,6 @@ Imports:
     DBI,
     duckdb,
     httr,
-    R.utils,
     tibble,
     purrr,
     sf,

[…]
