+++ NEWS.md
@@ -1,3 +1,8 @@
+# Version 0.4
+
+- Works with duckdb 0.3.4
+- Removes the databases used with older censo2017 versions
+
 # Version 0.3
 
 - Moves local database location according to CRAN request
