### 2021-02-05 50c7b71
using duckdb::dbSendQuery
