+++ R/download.R
@@ -21,6 +21,7 @@
 #' \dontrun{
 #' censo_descargar_base()
 #' }
+#' }
 censo_descargar_base <- function(ver = NULL, destdir = tempdir()) {
   msg("Descargando la base de datos desde GitHub...")
   
@@ -82,9 +83,9 @@ censo_descargar_base <- function(ver = NULL, destdir = tempdir()) {
     invisible(gc())
   }
   
-  metadatos <- data.frame(version_rsqlite = utils::packageVersion("RSQlite"),
+  metadatos <- data.frame(version_duckdb = utils::packageVersion("duckdb"),
                           fecha_modificacion = Sys.time())
-  metadatos$version_rsqlite <- as.character(metadatos$version_rsqlite)
+  metadatos$version_duckdb <- as.character(metadatos$version_duckdb)
   metadatos$fecha_modificacion <- as.character(metadatos$fecha_modificacion)
   
   con <- censo_bbdd()
