+++ R/schema.R
@@ -1,9 +1,9 @@
 create_schema <- function(con = censo_bbdd()) {
   # comunas ----
   
-  duckdb::dbGetQuery(con, "DROP TABLE IF EXISTS comunas")
+  DBI::dbGetQuery(con, "DROP TABLE IF EXISTS comunas")
   
-  duckdb::dbGetQuery(
+  DBI::dbGetQuery(
     con,
     "CREATE TABLE comunas (
 	comuna_ref_id INTEGER NOT NULL,
@@ -15,9 +15,9 @@ create_schema <- function(con = censo_bbdd()) {
   
   # hogares ----
   
-  duckdb::dbGetQuery(con, "DROP TABLE IF EXISTS hogares")
+  DBI::dbGetQuery(con, "DROP TABLE IF EXISTS hogares")
   
-  duckdb::dbGetQuery(
+  DBI::dbGetQuery(
     con,
     "CREATE TABLE hogares (
 	hogar_ref_id INTEGER NOT NULL,
@@ -38,9 +38,9 @@ create_schema <- function(con = censo_bbdd()) {
   
   # mapa comunas ----
   
-  duckdb::dbGetQuery(con, "DROP TABLE IF EXISTS mapa_comunas")
+  DBI::dbGetQuery(con, "DROP TABLE IF EXISTS mapa_comunas")
   
-  duckdb::dbGetQuery(
+  DBI::dbGetQuery(
     con,
     "CREATE TABLE mapa_comunas (
 	region VARCHAR(2) NULL,
@@ -51,9 +51,9 @@ create_schema <- function(con = censo_bbdd()) {
   
   # mapa provincias ----
   
-  duckdb::dbGetQuery(con, "DROP TABLE IF EXISTS mapa_provincias")
+  DBI::dbGetQuery(con, "DROP TABLE IF EXISTS mapa_provincias")
   
-  duckdb::dbGetQuery(
+  DBI::dbGetQuery(
     con,
     "CREATE TABLE mapa_provincias (
 	region VARCHAR(2) NULL,
@@ -63,9 +63,9 @@ create_schema <- function(con = censo_bbdd()) {
   
   # mapa regiones ----
   
-  duckdb::dbGetQuery(con, "DROP TABLE IF EXISTS mapa_regiones")
+  DBI::dbGetQuery(con, "DROP TABLE IF EXISTS mapa_regiones")
   
-  duckdb::dbGetQuery(
+  DBI::dbGetQuery(
     con,
     "CREATE TABLE mapa_regiones (
 	region VARCHAR(2) NOT NULL,
@@ -73,9 +73,9 @@ create_schema <- function(con = censo_bbdd()) {
   )
   
   # mapa zonas ----
-  duckdb::dbGetQuery(con, "DROP TABLE IF EXISTS mapa_zonas")
+  DBI::dbGetQuery(con, "DROP TABLE IF EXISTS mapa_zonas")
   
-  duckdb::dbGetQuery(
+  DBI::dbGetQuery(
     con,
     "CREATE TABLE mapa_zonas (
 	region VARCHAR(2) NULL,
@@ -87,9 +87,9 @@ create_schema <- function(con = censo_bbdd()) {
   
   # personas ----
   
-  duckdb::dbGetQuery(con, "DROP TABLE IF EXISTS personas")
+  DBI::dbGetQuery(con, "DROP TABLE IF EXISTS personas")
   
-  duckdb::dbGetQuery(
+  DBI::dbGetQuery(
     con,
     "CREATE TABLE personas (
 	persona_ref_id DOUBLE NULL,
@@ -132,9 +132,9 @@ create_schema <- function(con = censo_bbdd()) {
   
[…]
