### 2021-02-05 90e963c
fixes tests and removes metadatos table
