+++ R/schema.R
@@ -1,201 +1,0 @@
-create_schema <- function(con = censo_bbdd()) {
-  # comunas ----
-  
-  DBI::dbGetQuery(con, "DROP TABLE IF EXISTS comunas")
-  
-  DBI::dbGetQuery(
-    con,
-    "CREATE TABLE comunas (
-	comuna_ref_id INTEGER NOT NULL,
-	provincia_ref_id INTEGER NULL,
-	idcomuna VARCHAR NULL,
-	redcoden VARCHAR(5) NOT NULL,
-	nom_comuna VARCHAR NULL)"
-  )
-  
-  # hogares ----
-  
-  DBI::dbGetQuery(con, "DROP TABLE IF EXISTS hogares")
-  
-  DBI::dbGetQuery(
-    con,
-    "CREATE TABLE hogares (
-	hogar_ref_id INTEGER NOT NULL,
-	vivienda_ref_id INTEGER NULL,
-	nhogar INTEGER NULL,
-	tipo_hogar INTEGER NULL,
-	ncu_yern_nuer INTEGER NULL,
-	n_herm_cun INTEGER NULL,
-	nuc_herm_cun INTEGER NULL,
-	num_sueg_pad_abu INTEGER NULL,
-	nuc_pad_sueg_abu INTEGER NULL,
-	num_otros INTEGER NULL,
-	nuc_otros INTEGER NULL,
-	num_no_par INTEGER NULL,
-	nuc_no_par INTEGER NULL,
-	tot_nucleos INTEGER NULL)"
-  )
-  
-  # mapa comunas ----
-  
-  DBI::dbGetQuery(con, "DROP TABLE IF EXISTS mapa_comunas")
-  
-  DBI::dbGetQuery(
-    con,
-    "CREATE TABLE mapa_comunas (
-	region VARCHAR(2) NULL,
-	provincia VARCHAR(3) NULL,
-	comuna VARCHAR(5) NOT NULL,
-  geometry VARCHAR NULL)"
-  )
-  
-  # mapa provincias ----
-  
-  DBI::dbGetQuery(con, "DROP TABLE IF EXISTS mapa_provincias")
-  
-  DBI::dbGetQuery(
-    con,
-    "CREATE TABLE mapa_provincias (
-	region VARCHAR(2) NULL,
-	provincia VARCHAR(3) NOT NULL,
-  geometry VARCHAR NULL)"
-  )
-  
-  # mapa regiones ----
-  
-  DBI::dbGetQuery(con, "DROP TABLE IF EXISTS mapa_regiones")
-  
-  DBI::dbGetQuery(
-    con,
-    "CREATE TABLE mapa_regiones (
-	region VARCHAR(2) NOT NULL,
-  geometry VARCHAR NULL);"
-  )
-  
-  # mapa zonas ----
-  DBI::dbGetQuery(con, "DROP TABLE IF EXISTS mapa_zonas")
-  
-  DBI::dbGetQuery(
-    con,
-    "CREATE TABLE mapa_zonas (
-	region VARCHAR(2) NULL,
-	provincia VARCHAR(3) NULL,
-	comuna VARCHAR(5) NULL,
-	geocodigo VARCHAR(11) NOT NULL,
-  geometry VARCHAR NULL)"
-  )
-  
-  # personas ----
-  
-  DBI::dbGetQuery(con, "DROP TABLE IF EXISTS personas")
-  
-  DBI::dbGetQuery(
-    con,
-    "CREATE TABLE personas (
-	persona_ref_id DOUBLE NULL,
-	hogar_ref_id INTEGER NULL,
-	personan INTEGER NULL,
[…]
