+++ R/download.R
@@ -1,17 +1,12 @@
 #' Descarga la Base de Datos del Censo a tu Computador
 #'
-#' Este comando descarga la base de datos completa como un unico archivo zip que
-#' se descomprime para crear la base de datos local. La descarga es 
-#' 570 MB y la base de datos usa 8 GB en disco. Si no quieres descargar la base 
-#' de datos en tu home, ejecuta usethis::edit_r_environ() para crear la variable
-#' de entorno CENSO_BBDD_DIR con la ruta.
+#' Este comando descarga la base de datos completa como un unico archivo bz2 que
+#' se descomprime para dejar disponible la base de datos local en formato 
+#' SQLite.
 #'
 #' @param ver La version a descargar. Por defecto es la ultima version 
 #' disponible en GitHub. Se pueden ver todas las versiones en
 #' <https://github.com/pachamaltese/censo2017/releases>.
-#' 
-#' @param destdir La carpeta donde se va a descargar y descomprimir el archivo
-#' zip para crear la base de datos. Por defecto es una carpeta temporal.
 #'
 #' @return NULL
 #' @export
@@ -22,7 +17,7 @@
 #' censo_descargar_base()
 #' }
 #' }
-censo_descargar_base <- function(ver = NULL, destdir = tempdir()) {
+censo_descargar_base <- function(ver = NULL) {
   msg("Descargando la base de datos desde GitHub...")
   
   dir <- censo_path()
@@ -30,78 +25,23 @@ censo_descargar_base <- function(ver = NULL, destdir = tempdir()) {
   
   zfile <- get_gh_release_file("pachamaltese/censo2017",
                                tag_name = ver,
-                               dir = destdir
+                               dir = dir
   )
   ver <- attr(zfile, "ver")
   
   msg("Descomprimiendo la base de datos local...")
   
+  dfile <- gsub("\\.bz2", "", zfile)
   suppressWarnings(try(censo_desconectar_base()))
-  try(censo_borrar_base())
-  
-  utils::unzip(zfile, overwrite = TRUE, exdir = destdir)
-  
-  finp_tsv <- list.files(destdir, full.names = TRUE, pattern = "tsv")
-  finp_shp <- list.files(destdir, full.names = TRUE, pattern = "shp")
-  
-  invisible(create_schema())
-  
-  for (x in seq_along(finp_tsv)) {
-    tout <- gsub(".*/", "", gsub("\\.tsv", "", finp_tsv[x]))
-    
-    msg(sprintf("Creando tabla %s ...", tout))
-    
-    con <- censo_bbdd()
-    
-    suppressMessages(
-      DBI::dbExecute(
-        con,
-        paste0(
-          "COPY ", tout, " FROM '",
-          finp_tsv[x],
-          "' ( DELIMITER '\t', HEADER 1, NULL 'NA' )"
-        )
-      )
-    )
-    
-    DBI::dbDisconnect(con, shutdown = TRUE)
-    invisible(gc())
-  }
-  
-  # for (x in seq_along(finp_shp)) {
-  #   tout <- gsub(".*/", "", gsub("\\.shp", "", finp_shp[x]))
-  #   
-  #   msg(sprintf("Creando tabla %s ...", tout))
-  #   
-  #   con <- censo_bbdd()
-  #   
-  #   d <- sf::st_read(finp_shp[x], quiet = TRUE)
-  #   suppressMessages(DBI::dbWriteTable(con, tout, d, append = T, temporary = F))
-  #   
-  #   DBI::dbDisconnect(con, shutdown = TRUE)
-  #   rm(d)
-  #   invisible(gc())
-  # }
-  
[…]
