### 2021-02-05 ae8949b
trying a sustainable approach to use duckdb, rsqlite so slow
