+++ R/github.R
@@ -1,0 +1,38 @@
+#' @importFrom httr GET stop_for_status content accept write_disk progress
+#' @importFrom purrr keep
+get_gh_release_file <- function(repo, tag_name = NULL, dir = tempdir(),
+                                overwrite = TRUE) {
+  releases <- GET(
+    paste0("https://api.github.com/repos/", repo, "/releases")
+  )
+  stop_for_status(releases, "buscando versiones")
+  
+  releases <- content(releases)
+  
+  if (is.null(tag_name)) {
+    release_obj <- releases[1]
+  } else {
+    release_obj <- purrr::keep(releases, function(x) x$tag_name == tag_name)
+  }
+  
+  if (!length(release_obj)) stop("No se encuenta una version disponible \"",
+                                 tag_name, "\"")
+  
+  if (release_obj[[1]]$prerelease) {
+    msg("Estos datos aun no se han validado.")
+  }
+  
+  download_url <- release_obj[[1]]$assets[[1]]$url
+  filename <- basename(release_obj[[1]]$assets[[1]]$browser_download_url)
+  out_path <- normalizePath(file.path(dir, filename), mustWork = FALSE)
+  response <- GET(
+    download_url,
+    accept("application/octet-stream"),
+    write_disk(path = out_path, overwrite = overwrite),
+    progress()
+  )
+  stop_for_status(response, "downloading data")
+  
+  attr(out_path, "ver") <- release_obj[[1]]$tag_name
+  return(out_path)
+}
