+++ R/connect.R
@@ -1,5 +1,4 @@
 censo_path <- function() {
-  rsqlite_version <- utils::packageVersion("RSQLite")
   sys_censo_path <- Sys.getenv("CENSO_BBDD_DIR")
   sys_censo_path <- gsub("\\\\", "/", sys_censo_path)
   if (sys_censo_path == "") {
@@ -19,7 +18,7 @@ censo_check_status <- function() {
 #' Conexion a la Base de Datos del Censo
 #'
 #' Devuelve una conexion a la base de datos local. Esto corresponde a una
-#' conexion a una base SQLite compatible con DBI. A diferencia de
+#' conexion a una base DuckDB compatible con DBI. A diferencia de
 #' [censo2017::censo_tabla()], esta funcion es mas flexible y se puede usar con
 #' dbplyr para leer unicamente lo que se necesita o directamente con DBI para
 #' usar comandos SQL.
@@ -40,7 +39,11 @@ censo_check_status <- function() {
 #'  )
 #' }
 censo_bbdd <- function(dir = censo_path()) {
+  duckdb_version <- utils::packageVersion("duckdb")
+  db_file <- paste0(dir, "/censo2017_duckdb_v", gsub("\\.", "", duckdb_version), ".duckdb")
+  
   db <- mget("censo_bbdd", envir = censo_cache, ifnotfound = NA)[[1]]
+  
   if (inherits(db, "DBIConnection")) {
     if (DBI::dbIsValid(db)) {
       return(db)
@@ -51,8 +54,8 @@ censo_bbdd <- function(dir = censo_path()) {
 
   tryCatch({
     db <- DBI::dbConnect(
-      RSQLite::SQLite(),
-      paste0(dir, "/censo2017.sqlite")
+      duckdb::duckdb(),
+      db_file
     )
   },
   error = function(e) {
@@ -146,8 +149,7 @@ censo_estado <- function(msg = TRUE) {
     out <- TRUE
   } else {
     status_msg <- crayon::red(paste(cli::symbol$cross,
-    "La base de datos local del Censo 2017 esta vacia o daniada.
-  Descargala con censo_descargar_base()."))
+    "La base de datos local del Censo 2017 esta vacia, daniada o no es compatible con tu version de duckdb. Descargala con censo_descargar_base()."))
     out <- FALSE
   }
   if (msg) msg(status_msg)
@@ -157,7 +159,7 @@ censo_estado <- function(msg = TRUE) {
 censo_tables <- function() {
   c("comunas", "hogares", "mapa_comunas", "mapa_provincias",
     "mapa_regiones", "mapa_zonas", "personas", "provincias",
-    "regiones", "viviendas", "zonas")
+    "regiones", "viviendas", "zonas", "metadatos")
 }
 
 censo_cache <- new.env()
