+++ man/censo_descargar_base.Rd
@@ -12,9 +12,11 @@ disponible en GitHub. Se pueden ver todas las versiones en
 \url{https://github.com/pachamaltese/censo2017/releases}.}
 }
 \description{
-Este comando descarga la base de datos completa como un unico archivo bz2 que
-se descomprime para dejar disponible la base de datos local en formato
-SQLite.
+Este comando descarga la base de datos completa como un unico archivo zip que
+se descomprime para crear la base de datos local. La descarga es
+570 MB y la base de datos usa 8 GB en disco. Si no quieres descargar la base
+de datos en tu home, ejecuta usethis::edit_r_environ() para crear la variable
+de entorno CENSO_BBDD_DIR con la ruta.
 }
 \examples{
 \donttest{
