+++ NAMESPACE
@@ -7,10 +7,3 @@ export(censo_desconectar_base)
 export(censo_estado)
 export(censo_panel)
 export(censo_tabla)
-importFrom(httr,GET)
-importFrom(httr,accept)
-importFrom(httr,content)
-importFrom(httr,progress)
-importFrom(httr,stop_for_status)
-importFrom(httr,write_disk)
-importFrom(purrr,keep)

[…]
