+++ R/download.R
@@ -100,3 +100,40 @@ censo_descargar_base <- function(ver = NULL) {
   censo_panel()
   censo_estado()
 }
+
+get_gh_release_file <- function(repo, tag_name = NULL, dir = tempdir(),
+                                overwrite = TRUE) {
+  releases <- httr::GET(
+    paste0("https://api.github.com/repos/", repo, "/releases")
+  )
+  httr::stop_for_status(releases, "buscando versiones")
+  
+  releases <- httr::content(releases)
+  
+  if (is.null(tag_name)) {
+    release_obj <- releases[1]
+  } else {
+    release_obj <- purrr::keep(releases, function(x) x$tag_name == tag_name)
+  }
+  
+  if (!length(release_obj)) stop("No se encuenta una version disponible \"",
+                                 tag_name, "\"")
+  
+  if (release_obj[[1]]$prerelease) {
+    msg("Estos datos aun no se han validado.")
+  }
+  
+  download_url <- release_obj[[1]]$assets[[1]]$url
+  filename <- basename(release_obj[[1]]$assets[[1]]$browser_download_url)
+  out_path <- normalizePath(file.path(dir, filename), mustWork = FALSE)
+  response <- httr::GET(
+    download_url,
+    httr::accept("application/octet-stream"),
+    httr::write_disk(path = out_path, overwrite = overwrite),
+    httr::progress()
+  )
+  httr::stop_for_status(response, "downloading data")
+  
+  attr(out_path, "ver") <- release_obj[[1]]$tag_name
+  return(out_path)
+}
