+++ R/schema.R
@@ -40,54 +40,54 @@ create_schema <- function() {
 	tot_nucleos INTEGER NULL)"
   )
   
-  # mapa comunas ----
-  
-  DBI::dbSendQuery(con, "DROP TABLE IF EXISTS mapa_comunas")
-  
-  DBI::dbSendQuery(
-    con,
-    "CREATE TABLE mapa_comunas (
-	region VARCHAR(2) NULL,
-	provincia VARCHAR(3) NULL,
-	comuna VARCHAR(5) NOT NULL,
-  geometry VARCHAR NULL)"
-  )
-  
-  # mapa provincias ----
-  
-  DBI::dbSendQuery(con, "DROP TABLE IF EXISTS mapa_provincias")
-  
-  DBI::dbSendQuery(
-    con,
-    "CREATE TABLE mapa_provincias (
-	region VARCHAR(2) NULL,
-	provincia VARCHAR(3) NOT NULL,
-  geometry VARCHAR NULL)"
-  )
-  
-  # mapa regiones ----
-  
-  DBI::dbSendQuery(con, "DROP TABLE IF EXISTS mapa_regiones")
-  
-  DBI::dbSendQuery(
-    con,
-    "CREATE TABLE mapa_regiones (
-	region VARCHAR(2) NOT NULL,
-  geometry VARCHAR NULL);"
-  )
-  
-  # mapa zonas ----
-  DBI::dbSendQuery(con, "DROP TABLE IF EXISTS mapa_zonas")
-  
-  DBI::dbSendQuery(
-    con,
-    "CREATE TABLE mapa_zonas (
-	region VARCHAR(2) NULL,
-	provincia VARCHAR(3) NULL,
-	comuna VARCHAR(5) NULL,
-	geocodigo VARCHAR(11) NOT NULL,
-  geometry VARCHAR NULL)"
-  )
+#   # mapa comunas ----
+#   
+#   DBI::dbSendQuery(con, "DROP TABLE IF EXISTS mapa_comunas")
+#   
+#   DBI::dbSendQuery(
+#     con,
+#     "CREATE TABLE mapa_comunas (
+# 	region VARCHAR(2) NULL,
+# 	provincia VARCHAR(3) NULL,
+# 	comuna VARCHAR(5) NOT NULL,
+#   geometry VARCHAR NULL)"
+#   )
+#   
+#   # mapa provincias ----
+#   
+#   DBI::dbSendQuery(con, "DROP TABLE IF EXISTS mapa_provincias")
+#   
+#   DBI::dbSendQuery(
+#     con,
+#     "CREATE TABLE mapa_provincias (
+# 	region VARCHAR(2) NULL,
+# 	provincia VARCHAR(3) NOT NULL,
+#   geometry VARCHAR NULL)"
+#   )
+#   
+#   # mapa regiones ----
+#   
+#   DBI::dbSendQuery(con, "DROP TABLE IF EXISTS mapa_regiones")
+#   
+#   DBI::dbSendQuery(
+#     con,
+#     "CREATE TABLE mapa_regiones (
+# 	region VARCHAR(2) NOT NULL,
+#   geometry VARCHAR NULL);"
+#   )
+#   
+#   # mapa zonas ----
+#   DBI::dbSendQuery(con, "DROP TABLE IF EXISTS mapa_zonas")
+#   
+#   DBI::dbSendQuery(
+#     con,
+#     "CREATE TABLE mapa_zonas (
+# 	region VARCHAR(2) NULL,
+# 	provincia VARCHAR(3) NULL,
+# 	comuna VARCHAR(5) NULL,
+# 	geocodigo VARCHAR(11) NOT NULL,
+#   geometry VARCHAR NULL)"
+#   )
   
   # personas ----
   
[…]
