+++ R/download.R
@@ -1,10 +1,9 @@
 #' Descarga la Base de Datos del Censo a tu Computador
 #'
 #' Este comando descarga la base de datos completa como un unico archivo zip que
-#' se descomprime para crear la base de datos local. La descarga necesita 310 MB 
-#' libres y la base de datos 8 GB. Si no quieres descargar la base 
-#' de datos en tu home, ejecuta usethis::edit_r_environ() para crear la variable
-#' de entorno CENSO_BBDD_DIR con la ruta.
+#' se descomprime para crear la base de datos local. Si no quieres descargar la 
+#' base de datos en tu home, ejecuta usethis::edit_r_environ() para crear la 
+#' variable de entorno CENSO_BBDD_DIR con la ruta.
 #'
 #' @param ver La version a descargar. Por defecto es la ultima version 
 #' disponible en GitHub. Se pueden ver todas las versiones en
