+++ man/censo_descargar_base.Rd
@@ -13,10 +13,9 @@ disponible en GitHub. Se pueden ver todas las versiones en
 }
 \description{
 Este comando descarga la base de datos completa como un unico archivo zip que
-se descomprime para crear la base de datos local. La descarga necesita 310 MB
-libres y la base de datos 8 GB. Si no quieres descargar la base
-de datos en tu home, ejecuta usethis::edit_r_environ() para crear la variable
-de entorno CENSO_BBDD_DIR con la ruta.
+se descomprime para crear la base de datos local. Si no quieres descargar la
+base de datos en tu home, ejecuta usethis::edit_r_environ() para crear la
+variable de entorno CENSO_BBDD_DIR con la ruta.
 }
 \examples{
 \donttest{
