### 2021-02-07 #414 @mpaulacaldas
Estimados @pachamaltese @FvD @melvidoni. Encontrarán aquí mi revisión. Por favor no duden si tienen dudas.

## Revisión de un paquete

*Por favor trata de marcar tantas casillas como te sea posible y elabora tus argumentos en comentarios abajo de cada una. Tu revisión no esta limitada a estos temas, tal como se describe en la guia para revisores (Reviewer Guide)*

**Por favor describe cualquier relación de trabajo que tengas/hayas tenido con los autores del paquete)**

- [x] Como revisor confirmo que no tengo [conflictos de interés](https://devguide.ropensci.org/policies.html#coi) para poder hacer la revisión de este trabajo (si no estas segura si tienes un conflicto por favor entra en contacto con tu editor _antes_ de arrancar con la revisión.

#### Documentación

El paquete incluye todos los siguiente tipos de documentación:

- [x] **Una declaración de necesidades** que claramente describe las necesidades que el software esta diseñado a resolver y el publico meta que busca atender en el archivo README

  > El README describe el objetivo principal de manera breve: el paquete provee
  > acceso conveniente al censo 2017, que tiene más de 17 millones de registros.
  >
  > Dejo aquí, y en la sección siguiente ciertos comentarios adicionales que
  > creo ayudarían a mejorar la calidad del README:
  >
  > - Sería útil incluir un vínculo a la página principal o a la documentación
  >   oficial del censo.
  > - El README da a entender que el paquete provee shapefiles, pero ninguna de
  >   las tablas de la base tiene información de geometrías. Si `censo2017`
  >   se limita a las tablas, y `chilemapas` al aspecto geográfico, debería
  >   estar más clara la distinción en el README.
  > - No queda claro a partir del README cuál es la principal ventaja de la base
  >   SQLite (o es ahora duckdb?) con respecto a la base PostGIS. Sería bueno
  >   mencionarlo en una frase, dado que el blog post asociado está en inglés.
  > - La falta de tildes/acentos en el README dificulta la lectura, y puede
  >   prestarse a confuciones, por ejemplo, respecto al nombre de regiones (e.g.
  >   Niuble versus Ñuble). Si esto es necesario (según lo que veo en el archivo
  >   DESCRIPTION), debería mencionarse en el README.

- [ ] **Instrucciones de instalación** de la versión en desarrollo del paquete incluyendo cualquier dependencia no-estándar en el archivo README

  > Dos sugerencias, dado el tamaño del paquete:
  >
  > - Debería darse información sobre los requisitos computacionales del paquete
  >   a nivel del README, similar a lo que se hace en el mensaje de `.onAttach()`.
  > - Puede ser también una buena idea explicar en el README que después de la
  >   primera llamada a `library(censo2017)` se le pedirá al usuario que
  >   descargue la base usando `censo_descargar_base()`, y que puede modificar
  >   la ruta de descarga con la variable de entorno `CENSO_BBDD_DIR`.

- [ ] **Viñeta(s)** demostrando la funcionalidad principal que ademas corren localmente

  > La viñeta no está disponible cuando se descarga el paquete, incluso usando
  > `remotes::install_github("pachamaltese/censo2017", build_vignettes = TRUE)`.
  > Tampoco se encuentra disponible cuando hago el build de manera local.
  >
  > En términos de contenidos, la viñeta presenta un buen ejemplo del tipo de
  > exploración que haría un usuario del paquete.
  >
  > Un comentario menor: Los mensajes descritos en "Notas" llegan un poco tarde,
  > y podrían ser incorporados antes en el texto. Por ejemplo, se podría
  > explicar desde antes que `tbl()` es necesario para abrir la connexión a la
  > base, o que el paquete `chilemapas` es necesario para la generación del
  > gráfico.

- [ ] **Documentación de las funciones** exportadas

  > `censo_bbdd()` menciona que tiene como default a `censo_path()`, que no
  > está exportada. Creo que sería buena idea exportarla, ya que le permitiría
  > al usuario verificar la ruta donde está guardada la base, y facilita el
  > mantenimiento de la documentación (ver siguiente nota).
  >
  > `censo_descargar_base()` dice que la base se descarga por default al home,
  > pero en este issue comentas que cambiaste el default debido a políticas de
  > CRAN. Faltaría poner al día esa información en la documentación. Una
  > solución que requeriría baja inversión sería mencionar que la ruta se puede
  > consultar usando `censo_path()`.

- [x] **Ejemplos** (que corren localmente) para todas las funciones exportadas
- [ ] **Directrices comunitarias** incluyendo una guia de contribución en el archivo README o el archivo CONTRIBUTING y un archivo DESCRIPTION que incluye `URL`, `BugReports` and `Maintainer` (todas en inglés por concenvión y para que puedan ser autogeneradas con `Authors@R`).

  > Falta guia de contribución en el README y/o archivo CONTRIBUTING.

#### Funcionalidad

- [x] **Instalación:** La instalación se completa con éxito tal como fue documentada.

  > Respecto al mensaje que sugiere llamar a  `censo_descargar_base()`, me
  > parece que sería bueno que mencione también que la ruta de descarga se
  > puede configurar usando la variable de entorno `CENSO_BBDD_DIR` como lo
  > indicas en la documentación. Como usuaria, mi primera pregunta fue, ¿en
  > dónde se están guardando estas 6.5GB? ¿Puedo configurar la ruta? ¿Qué hago
  > una vez se acabe esta revisión para borrarla?

- [ ] **Funcionalidad:** Toda afirmación de funcionalidad del software se confirma como existente.

  > `censo_borrar_base()` no elimina el directorio especificado en
  > `censo_path()`, solo desconecta la base. El test correspondiente no verifica
  > que el directorio y sus contenidos hayan sido eliminados del disco.

- [x] **Desempeño:** Toda afirmación de desempeño del software se confirma como alcanzada.
- [x] **Pruebas automáticas:** Hay pruebas unitarias que cubren las funciones esenciales dentro del paquete con un rango razonable de entradas y condiciones. Todas las pruebas corren en la maquina local.
- [ ] **Directrices de empaque**: El paquete cumple con las directrices de empaque de rOpenSci.

  > Dejo aquí los elementos mencionados en [las directrices de empaque](https://devguide.ropensci.org/building.html#package-name-and-metadata)
  > que considero que faltan por implementar:
  >
  > - [ ] Se incluye un README, con el formato recomendado -- *Falta link a la
  >       viñeta o breve demostración (según criterio de multiples puntos de
  >       entrada). También sugiero explicar de mejor manera la interacción
  >       con `chilemapas`, y los requisitos de instalación, como menciono en
  >       mis comentarios anteriores.*
  > - [ ] `DESCRIPTION` contiene información de la organización responsable por
  >       proveer los datos y un URL a la página que provee, describe o permite
  >       el acceso a los datos -- *Se menciona al INE, pero no hay vínculo.*
  > - [ ] Se usan los mensajes de inicio de manera estrictamente necesaria --
  >       *Como recomiendan las guidelines, recomendaría quitar el mensaje de cómo
  >       citar. Recomiendo también cambiar el mensaje que le pide al usuario leer
  >       la documentación (más detalles abajo).*

Estimación de horas dedicadas a la revisión: 5

- [x] Si la o las persona(s) autora(s) lo considera(n) apropiado, yo estoy de acuerdo con que me reconozcan como revisor del paquete (el rol "rev') en la el archivo DESCRIPTION del paquete.

---

### Comentarios de la revisión

Muchas gracias @pachamaltese por esta excelente contribución. Como menciona @FvD, este paquete puede llegar a facilitar de manera importante el trabajo de investigadores, estudiantes, funcionarios públicos, periodistas, etc. Como lo veo, la mayor  ventaja para los futuros usuarios es que pueden tener acceso a datos  voluminosos por medio de R, sin tener que aprender a manejar otro tipo de tecnologías.

Dado que el tipo de usuarios de este paquete puede incluir a bastantes principiantes, entiendo que es imperativo tener una documentación y una viñeta clara. Creo que en su estado la documentación del paquete es muy buena y llega a cumplir 90% de este objetivo. Mis sugerencias se concentran en el 10% de detalles que considero quedan faltando.

Finalmente, algo que saltó a mi vista es el mensaje inicio cuando se carga el paquete, donde le pides a los usuarios que por favor lean la documentación. Este mensaje no es informativo para principiantes (porque justamente puede que no sepan cómo acceder a la documentación), y corre el riesgo de desanimar a potenciales usuarios. Si decides guardar el mensaje, sugeriría dar un poco más de ayuda al usuario, por ejemplo, indicando cómo acceder a la vignette, o dirigiendólo a la documentación en línea. Podría leer, por ejemplo:

  ```
  Por favor, lee la documentacion. Un buen lugar para empezar es `vignette("censo2017")`.
  ```

O de manera alternativa:

  ```
  La documentación del paquete y ejemplos de uso se encuentran en https://pacha.dev/censo2017/.
  ```

Como mencioné, noté también la falta de tíldes en la documentación. ¿Fue esto una decisión tomada debido a una sugerencia de CRAN? Esta pregunta la hago es por curiosidad propia. Si ese fue el caso, creo que sería algo útil para incluir en una eventual lista de [CRAN 'gotchas'](https://devguide.ropensci.org/building.html#crangotchas) de paquetes en otras lenguas.
