### 2021-02-09 #414 @pachamaltese
@mpaulacaldas @FvD muchas gracias por sus tremendos aportes !!

La semana pasada incorporé algunas de las cosas que me han mencionado y en verdad fue muy difícil hacer este paquete por tres principales razones:

1. el censo original viene en formato REDATAM, por lo que solo se puede leer en Windows y hay que usar un convertidor con el que llevé todo a PostreSQL y luego a una base local
2. ese archivo viene en un DVD, por lo que tuve que conseguir un lector externo
3. la documentación oficial es confusa, las cartografías no tienen los mismos campos entre regiones, hay que navegar muchos sitios, etc.

Cambié el mensaje de inicio, aún debo incluir la totalidad de lo que me comentan. Efectivamente las vignettes necesitan chilemapas, otro paquete que partió bajo la idea de "esto me está tomando mucho tiempo ¿quizá a cuántos más les pasa lo mismo?".

En la última iteración del paquete opté por sacar las cartografías y moverlas a otro repositorio (https://github.com/pachamaltese/cartografias-censo2017), lo cual reduce el peso de la base y así reduzco al mínimo las dependencias. De hecho chilemapas tiene toda una guía y one-liners de lo que hay que instalar para que funcione con sf y todas las dependencias.

En las distintas iteraciones migré de SQLite a DuckDB, que no es tan eficiente en uso de disco, pero es muchísimo más rápido y el último push se alimenta de archivos TSV que están en el release del repositorio, de modo que si el usuario tiene duckdb 0.2.1, no habrá problema, a diferencia de cuando yo subía un .bz2 que era específico para la versión de RSQLite o DuckDB. Ahora las tablas se crean al momento de ejecutar `censo_descargar_base()`.

Dejaré acá una lista de checks de todo lo que me han comentado:

- [ ] Agregar "La documentación del paquete y ejemplos de uso se encuentran en https://pacha.dev/censo2017/." al msje de inicio
- [x] Tildes en la documentación *escribi todo sin acentos, estilo los ninios y la ninias*
- [ ] README: link a la viñeta
- [ ] Explicar de mejor manera la interacción con chilemapas
- [ ] Explicar requisitos de instalación
- [ ] Links a los multiples sitios del INE
- [ ] Quitar el mensaje de cómo citar
- [ ] Cambiar el mensaje que le pide al usuario leer la documentación (más detalles abajo)
- [ ] `censo_borrar_base()` no elimina el directorio especificado en `censo_path()`
- [ ] El test correspondiente no verifica que el directorio y sus contenidos hayan sido eliminados del disco
- [ ] Falta guia de contribución en el README y/o archivo CONTRIBUTING.
- [x] Puede ser también una buena idea explicar en el README que después de la primera llamada a library(censo2017) se le pedirá al usuario que descargue la base usando censo_descargar_base(), y que puede modificar la ruta de descarga con la variable de entorno CENSO_BBDD_DIR. *Listo y con referencia a usethis::*
- [ ] Sacar referencias a los mapas, pues las cartografías las movi a otro repositorio
- [ ] Pulida a nivel de documentación
- [ ] Top level documentation
- [ ] Usar codemetar::
- [ ] Mensaje "Error, no esta disponible esta tabla, estas seguro que instalaste los datos con censo_descargar_base()`."
- [ ] Extender o los ejemplos o la descripción de algunas funciones
- [ ] Escribir propósito del paquete (para que usuario, etc)
- [ ] Detallar menciones a databases.pacha.dev

Por ahora, les dejo algunas interrogantes que me han surgido:
- [ ] ¿ven valor en escribir una función (tipo anexo) que descargue/lea las cartografías detalladas? esta podría no tener una dependencia explícita, sino usar un require y con base en ello dar mensajes del tipo "instala sf para poder usar estas cartografías". De todos modos, chilemapas ya incorpora cartografías simplificadas y en formato R
- [ ] ¿creen que es conveniente incluir la codificación de variables? la vinieta hace referencia a https://databases.pacha.dev/censo2017-descripcion-variables.xml y la verdad es que me he quebrado la cabeza intentando llevar eso a una tabla en formato tidy de la forma
```
tabla,columna,codigo,descripcion
vivienda,p01,1,casa
vivienda,p01,2,departamento
..., ..., ..., ...
