### 2021-02-28 #414 @pachamaltese
> Gracias @pachamaltese por el resumen de los cambios a venir. Para responder a tus preguntas:
>
> > ¿ven valor en escribir una función (tipo anexo) que descargue/lea las cartografías detalladas? esta podría no tener una dependencia explícita, sino usar un require y con base en ello dar mensajes del tipo "instala sf para poder usar estas cartografías". De todos modos, chilemapas ya incorpora cartografías simplificadas y en formato R
>
> Creo que esto en verdad depende del tipo de usuario. Me imagino que para muchos, las cartografías simplificadas de `chilemapas` bastarán, pero para otros (e.g. investigadores) puede que sea útil tener el las geometrías "detalladas" o "oficiales", en especial si estas muestran cambios en límites administrativos.
>
> Mencionas también que los datos no son fáciles de extraer y tratar (tuviste que leer el formato REDATAM y limpiar las cartografías). Si este trabajo ya está hecho y tomaría relativamente poco esfuerzo crear una función que descargue estos archivos, me parecería una buena adición.
>
> > ¿creen que es conveniente incluir la codificación de variables? la vinieta hace referencia a https://databases.pacha.dev/censo2017-descripcion-variables.xml y la verdad es que me he quebrado la cabeza intentando llevar eso a una tabla en formato tidy de la forma
>
> Incluir la codificación de las variables me parece importante, aunque no tiene necesariamente que venir en formato tidy si consideras que requeriría mucho trabajo limpiar el XML. ¿Existe un PDF o sitio web con la descripción de las variables? Se podría incluir un vínculo en la documentación del paquete (e.g. README, vignette, `?censo_tabla()`).

muchas gracias, incorporé varias de las sugerencias (me faltan algunas)
el principal problema con el REDATAM es que no es reproducible, quizá se podría portar el convertidor de C# a C++ y de allí usarlo con Rcpp.

te aviso en cuanto tenga todo
