### 2021-02-28 c181035
cambios pedidos rOpenSci; documentacion + contribuciones
