### 2021-02-28 #414 @mpaulacaldas
No entiendo tu comentario de REDATAM 😅 . Creo que en algún momento nos perdimos. En mi respuesta mencioné REDATAM solo como ejemplo. Discúlpa si lo hice sonar como si necesitaras hacer más trabajo.
