### 2021-03-01 cc83ce0
descripcion variables XML -> tibble, pasar a SQL?
