+++ data-raw/convertir-xml-a-tidy.R
@@ -1,0 +1,394 @@
+# I had to convert from XML to JSON in the terminal
+# sudo apt install libxml-compile-perl libjson-perl
+
+library(xml2)
+library(tidyverse)
+
+d <- read_xml("data-raw/censo2017-descripcion-variables.xml")
+
+cadena_a_titulo <- function(x) {
+  x %>% 
+    iconv(., to = "UTF-8") %>% 
+    str_trim() %>% 
+    ifelse(. == "", NA, .) %>% 
+    str_to_title() %>% 
+    str_replace_all(., "/A", "/a") %>% 
+    str_replace_all(., " De ", " de ") %>% 
+    str_replace_all(., " De\\)", " de)") %>% 
+    str_replace_all(., " Del ", " del ") %>% 
+    str_replace_all(., " O ", " o ") %>% 
+    str_replace_all(., " Y ", " y ") %>% 
+    str_replace_all(., " Por ", " por ") %>% 
+    str_replace_all(., " En ", " en ") %>% 
+    str_replace_all(., " U ", " u ") %>% 
+    str_replace_all(., " El ", " el ") %>% 
+    str_replace_all(., " La ", " la ") %>% 
+    str_replace_all(., " Al ", " al ") %>% 
+    str_replace_all(., "\\(Grupo", " (Grupo") %>% 
+    str_replace_all(., "_recode", " recodificado/a")
+}
+
+# explorar ----
+
+# persona <-
+#   xml_attrs(xml_child(xml_child(xml_child(
+#     xml_child(xml_child(xml_child(
+#       xml_child(xml_child(xml_child(xml_child(
+#         d, 2
+#       ), 2), 3), 3), 3
+#     ), 2), 3), 2
+#   ), 14), 13))
+# 
+# hogar <-
+#   xml_attrs(xml_child(xml_child(xml_child(
+#     xml_child(xml_child(xml_child(
+#       xml_child(xml_child(xml_child(d, 2), 2), 3), 3
+#     ), 3), 2), 3
+#   ), 2), 14))
+# 
+# vivienda <-
+#   xml_attrs(xml_child(xml_child(xml_child(
+#     xml_child(xml_child(xml_child(
+#       xml_child(xml_child(d, 2), 2), 3
+#     ), 3), 3), 2
+#   ), 3), 2))
+# 
+# zonaloc <-
+#   xml_attrs(xml_child(xml_child(xml_child(
+#     xml_child(xml_child(xml_child(xml_child(
+#       d, 2
+#     ), 2), 3), 3), 3
+#   ), 2), 3))
+# 
+# distrito <-
+#   xml_attrs(xml_child(xml_child(xml_child(
+#     xml_child(xml_child(d, 2), 2), 3
+#   ), 3), 3))
+# 
+# comuna <-
+#   xml_attrs(xml_child(xml_child(xml_child(xml_child(
+#     d, 2
+#   ), 2), 3), 3))
+# 
+# provincia <- xml_attrs(xml_child(xml_child(xml_child(d, 2), 2), 3))
+# 
+# region <- xml_attrs(xml_child(xml_child(d, 2), 2))
+# 
+# censo <- xml_attrs(xml_child(xml_child(d, 2), 2))
+
+# personas ----
+
+personas <-
+  xml_child(xml_child(xml_child(xml_child(
+    xml_child(xml_child(xml_child(
+      xml_child(xml_child(xml_child(d, 2), 2), 3), 3
+    ), 3), 2), 3
+  ), 2), 14), 13)
+
+# xml_attrs(xml_child(personas, 1))
+# xml_attrs(xml_child(personas, 34))
+personas2 <- xml_attrs(xml_children(personas))
+
+personas2 <- bind_rows(personas2) %>% 
[…]
