### 2021-03-01 9363a5c
tablas de variables a TSV + saca indices que no aportan
