### 2021-03-01 6e0e49e
cod. variables en sql
