+++ R/download.R
@@ -43,14 +43,12 @@ censo_descargar_base <- function(ver = NULL) {
   
   if (!any(grepl(db_pattern, existing_files))) {
     try(censo_borrar_base())
-    
   }
   
   utils::unzip(zfile, overwrite = TRUE, exdir = destdir)
   
   finp_tsv <- list.files(destdir, full.names = TRUE, pattern = "tsv")
-  # finp_shp <- list.files(destdir, full.names = TRUE, pattern = "shp")
-  
+
   invisible(create_schema())
   
   for (x in seq_along(finp_tsv)) {
@@ -74,21 +72,6 @@ censo_descargar_base <- function(ver = NULL) {
     DBI::dbDisconnect(con, shutdown = TRUE)
     invisible(gc())
   }
-  
-  # for (x in seq_along(finp_shp)) {
-  #   tout <- gsub(".*/", "", gsub("\\.shp", "", finp_shp[x]))
-  # 
-  #   msg(sprintf("Creando tabla %s ...", tout))
-  # 
-  #   con <- censo_bbdd()
-  # 
-  #   d <- sf::st_read(finp_shp[x], quiet = TRUE)
-  #   suppressMessages(DBI::dbWriteTable(con, tout, d, append = T, temporary = F))
-  # 
-  #   DBI::dbDisconnect(con, shutdown = TRUE)
-  #   rm(d)
-  #   invisible(gc())
-  # }
   
   metadatos <- data.frame(version_duckdb = utils::packageVersion("duckdb"),
                           fecha_modificacion = Sys.time())
