### 2021-03-01 360f268
agrega cod. variables a verificacion base local
