+++ R/utils.R
@@ -30,3 +30,14 @@ in_chk <- function() {
     )
   )
 }
+
+read_table_error <- function(e) {
+  e <- as.character(e)
+  # return(e)
+  msg <- c(
+    sprintf("No esta disponible la tabla %s.", get("tabla", env = 1)),
+    "\nVerifica que escribiste el nombre correctamente y que instalaste los",
+    "\ndatos con censo_descargar_base()."
+  )
+  stop(msg, call. = FALSE)
+}
