### 2021-04-03 c7f7711
elimina carpeta censo2017 de directorio de usuario
