+++ R/remove.R
@@ -15,5 +15,4 @@ censo_borrar_base <- function() {
   suppressWarnings(censo_desconectar_base())
   try(unlink(censo_path(), recursive = TRUE))
   update_censo_pane()
-  censo_panel()
 }
