### 2021-04-04 #414 @pachamaltese
hola @melvidoni @mpaulacaldas @FvD
disculpen la demora, hoy me he hecho el tiempo para revisar esto y tengo algunos cambios interesantes, el más relevante, creo, fue no depender del XML para la consulta de los significados de las variables (e.g. p01 significa "vivienda" y en p01 significa "apartamento")

antes: ir a databases.pacha.dev, buscar el censo y abrir el enlace al xml https://databases.pacha.dev/censo2017-descripcion-variables.xml

ahora:
```
library(censo2017)
library(dplyr)
# con la bbdd instalada
variables <- censo_tabla("variables")
variables_codificacion <- censo_tabla("variables_codificacion")

variables %>% filter(variable == "p01")
variables_codificacion %>% filter(variable == "p01")

# resultado
# A tibble: 1 x 5
  tabla     variable descripcion      tipo    rango
  <chr>     <chr>    <chr>            <chr>   <chr>
1 viviendas p01      Tipo de Vivienda integer 1 - 10

# A tibble: 12 x 4
   tabla     variable valor descripcion
   <chr>     <chr>    <int> <chr>
 1 viviendas p01          1 Casa
 2 viviendas p01          2 Departamento en Edificio
 3 viviendas p01          3 Vivienda Tradicional Indígena (Ruka, Pae Pae u Otras)
 4 viviendas p01          4 Pieza en Casa Antigua o en Conventillo
 5 viviendas p01          5 Mediagua, Mejora, Rancho o Choza
 6 viviendas p01          6 Móvil (Carpa, Casa Rodante o Similar)
 7 viviendas p01          7 Otro Tipo de Vivienda Particular
 8 viviendas p01          8 Vivienda Colectiva
 9 viviendas p01          9 Operativo Personas en Tránsito (No Es Vivienda)
10 viviendas p01         10 Operativo Calle (No Es Vivienda)
11 viviendas p01         11 Valor Perdido
12 viviendas p01          0 No Aplica
```

quitaré las referencias a databases.pacha.dev, es más, podría sacar el censo2017 de allí pues me resulta mucho mejor tener la versión tidy en local y dejar el servidor para las otras bases de datos que se utilizan en docencia en varias universidades
