+++ README.Rmd
@@ -1,0 +1,116 @@
+---
+output: github_document
+---
+
+<!-- README.md is generated from README.Rmd. Please edit that file -->
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.path = "man/figures/README-",
+  out.width = "100%"
+)
+```
+
+# Censo 2017 (Paquete R)
+
+<!-- badges: start -->
+[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://lifecycle.r-lib.org/articles/figures/lifecycle-stable.svg)](https://lifecycle.r-lib.org/articles/stages.html#stable-1)
+[![Lifecycle: stable](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#stable)
+[![GH-actions](https://github.com/pachamaltese/censo2017/workflows/R-CMD-check/badge.svg)](https://github.com/pachamaltese/censo2017/actions)
+[![codecov](https://codecov.io/gh/pachamaltese/censo2017/branch/main/graph/badge.svg?token=XI59cmGd15)](https://codecov.io/gh/pachamaltese/censo2017)
+[![CRAN status](https://www.r-pkg.org/badges/version/censo2017)](https://CRAN.R-project.org/package=censo2017)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4277761.svg)](https://doi.org/10.5281/zenodo.4277761)
+<!-- badges: end -->
+
+# Acerca de
+
+Provee un acceso conveniente a más de 17 millones de registros de la base de datos del Censo 2017. Los datos fueron importados desde el DVD oficial del INE usando el [Convertidor REDATAM](https://github.com/discontinuos/redatam-converter/) creado por Pablo De Grande y además se proporcionan los mapas que acompañan a estos datos. Estos mismos datos en DVD posteriormente quedaron disponibles en las [Bases de Datos del INE](https://www.ine.cl/estadisticas/sociales/censos-de-poblacion-y-vivienda/poblacion-y-vivienda).
+
+Después de la primera llamada a `library(censo2017)` se le pedirá al usuario que descargue la base usando `censo_descargar_base()` y se puede modificar la ruta de descarga con la variable de entorno `CENSO_BBDD_DIR`. La variable de entorno se puede crear con `usethis::edit_r_environ()`.
+
+El sitio de la documentación es pacha.dev/censo2017 y las viñetas con ejemplos están disponibles en pacha.dev/censo2017/articles.
+
+# Instalación
+
+Versión estable
+```
+install.packages("censo2017")
+```
+
+Versión de desarrollo
+```
+# install.packages("remotes")
+remotes::install_github("pachamaltese/censo2017")
+```
+
+# Valor agregado sobre los archivos SHP y REDATAM del INE
+
+Esta version de la base de datos del Censo 2017 presenta algunas diferencias respecto de la original que se obtiene en DVD y corresponde a una versión DuckDB derivada a partir de los Microdatos del Censo 2017 en formato DVD. 
+
+La modificacion sobre los archivos originales, que incluyen geometrias detalladas disponibles en [Cartografias Censo2017](https://github.com/pachamaltese/cartografias-censo2017), consistió en unir todos los archivos SHP regionales en una ínica tabla por nivel (e.g en lugar de proveer `R01_mapa_comunas`, ..., `R15_mapa_comunas` combiné las 15 regiones en una única tabla `mapa_comunas`).
+
+Los cambios concretos respecto de la base original son los siguientes:
+
[…]
