+++ README.md
@@ -1,85 +1,174 @@
+
+<!-- README.md is generated from README.Rmd. Please edit that file -->
+
+# Censo 2017 (Paquete R)
+
 <!-- badges: start -->
-[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://lifecycle.r-lib.org/articles/figures/lifecycle-stable.svg)](https://lifecycle.r-lib.org/articles/stages.html#stable-1)
-[![Lifecycle: stable](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#stable)
+
+[![Project Status: Active – The project has reached a stable, usable
+state and is being actively
+developed.](https://lifecycle.r-lib.org/articles/figures/lifecycle-stable.svg)](https://lifecycle.r-lib.org/articles/stages.html#stable-1)
+[![Lifecycle:
+stable](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#stable)
 [![GH-actions](https://github.com/pachamaltese/censo2017/workflows/R-CMD-check/badge.svg)](https://github.com/pachamaltese/censo2017/actions)
 [![codecov](https://codecov.io/gh/pachamaltese/censo2017/branch/main/graph/badge.svg?token=XI59cmGd15)](https://codecov.io/gh/pachamaltese/censo2017)
-[![CRAN status](https://www.r-pkg.org/badges/version/censo2017)](https://CRAN.R-project.org/package=censo2017)
+[![CRAN
+status](https://www.r-pkg.org/badges/version/censo2017)](https://CRAN.R-project.org/package=censo2017)
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4277761.svg)](https://doi.org/10.5281/zenodo.4277761)
 <!-- badges: end -->
 
 # Acerca de
 
-Provee un acceso conveniente a mas de 17 millones de registros de la base de datos del Censo 2017. Los datos fueron importados desde el DVD oficial del INE usando el Convertidor REDATAM creado por Pablo De Grande y ademas se proporcionan los mapas que acompanian a estos datos. Posteriormente estos datos quedaron disponibles en las [Bases de Datos del INE](https://www.ine.cl/estadisticas/sociales/censos-de-poblacion-y-vivienda/poblacion-y-vivienda).
+Provee un acceso conveniente a más de 17 millones de registros de la
+base de datos del Censo 2017. Los datos fueron importados desde el DVD
+oficial del INE usando el [Convertidor
+REDATAM](https://github.com/discontinuos/redatam-converter/) creado por
+Pablo De Grande y además se proporcionan los mapas que acompañan a estos
+datos. Estos mismos datos en DVD posteriormente quedaron disponibles en
+las [Bases de Datos del
+INE](https://www.ine.cl/estadisticas/sociales/censos-de-poblacion-y-vivienda/poblacion-y-vivienda).
 
-Despues de la primera llamada a `library(censo2017)` se le pedira al usuario que descargue la base usando `censo_descargar_base()` y se puede modificar la ruta de descarga con la variable de entorno `CENSO_BBDD_DIR`. La variable de entorno se puede crear con `usethis::edit_r_environ()`.
+Después de la primera llamada a `library(censo2017)` se le pedirá al
+usuario que descargue la base usando `censo_descargar_base()` y se puede
+modificar la ruta de descarga con la variable de entorno
+`CENSO_BBDD_DIR`. La variable de entorno se puede crear con
+`usethis::edit_r_environ()`.
 
-El sitio de la documentacion es pacha.dev/censo2017 y las vinietas estan disponibles en pacha.dev/censo2017/articles.
[…]
