+++ vignettes/censo2017.Rmd
@@ -18,16 +18,20 @@ knitr::opts_chunk$set(
 )
 ```
 
-# Introduccion
+# Introducción
 
-Este paquete se integra perfectamente con el tidyverse y se dara un ejemplo muy basico
-para mostrar las principales funciones.
+Este paquete se integra perfectamente con el tidyverse, con el que daremos un ejemplo muy básico para mostrar sus principales funciones.
 
-# Aproximacion de la poblacion con el grado de doctor en la Region del Bio Bio
+# Aproximación de la poblacion con el grado de doctor en la Región del Bio Bio.
 
-Se procedera a obtener una aproximacion usando dplyr ya que puede haber personas que no son de la comuna y aparecen censadas. Sin embargo no se hara el filtro que corrige esto para mantener el ejemplo lo mas simple posible.
+Se procederá a obtener una aproximación usando dplyr ya que, puede haber personas que no sean de la comuna y aparezcan censadas en ella. Sin embargo, cabe aclarar que para esta ocasión no haremos el filtro que corrige esto. Nuestra idea es mantener el ejemplo lo más simple posible.
 
-Primero se cargan los paquetes necesarios.
+Primero que todo, se cargan los paquetes necesarios.
+
+censo2017: Provee un acceso conveniente a mas de 17 millones de registros de la base de datos del Censo 2017.
+dplyr: Proporciona una forma bastante ágil de manejar los ficheros de datos de R. Además, el código escrito usando este paquete (especialmente usando la sintaxis en cadena que veremos más adelante) es mucho más legible y fácil de entender que el habitual.
+ggplot2: Provee una estructura para que podamos especificar qué variables graficar, cómo deben ser presentadas y otras propiedades visuales generales.
+chilemapas: Mapas terrestres con topologias simplificadas.
 ```{r, warning=FALSE, message=FALSE}
 library(censo2017)
 library(dplyr)
@@ -35,7 +39,9 @@ library(ggplot2)
 library(chilemapas)
 ```
 
-Hay que realizar algunos cruces de tablas de manera de filtrar la region que nos interesa. Comienzo con la tabla zonas, genero la provincia a partir del geocodigo y luego filtro para unir hasta llegar a la tabla personas. La variable `p15`, cuya descripcion esta en el [repositorio](https://github.com/pachamaltese/censo2017/blob/main/data-raw/censo2017-descripcion-variables.xml), que interesa para este ejemplo.
+Hay que realizar algunos cruces de tablas, de manera de filtrar la región que nos interesa.
+
+Comenzamos con la tabla zonas: generamos la provincia a partir del geocodigo y luego filtro para unir hasta llegar a la tabla personas. Nos interesa utilizar la variable `p15`, cuya descripción está en el [repositorio](https://github.com/pachamaltese/censo2017/blob/main/data-raw/censo2017-descripcion-variables.xml), para este ejemplo.
 ```{r, warning=FALSE, message=FALSE, eval=FALSE}
 nivel_educacional_biobio <- tbl(censo_bbdd(), "zonas") %>% 
   mutate(
@@ -50,7 +56,7 @@ nivel_educacional_biobio <- tbl(censo_bbdd(), "zonas") %>%
   collect()
 ```
 
-Con lo anterior se puede agrupar por o comuna y obtener la cuenta.
+Con lo anterior, los niveles educacionales de las personas censadas se pueden agrupar por comuna y obtener la cuenta proporcionada en base a la suma total.
 ```{r, warning=FALSE, message=FALSE, eval=FALSE}
 nivel_educacional_biobio <- nivel_educacional_biobio %>% 
   group_by(comuna, nivel_educ) %>%
[…]
