### 2021-04-10 #414 @mpaulacaldas
Hola a todos,

He vuelto a hacer una revisión del paquete y confirmo que @pachamaltese ha incorporado mis principales comentarios. Por lo tanto, considero que el paquete está listo para ser aprovado 🎉

Dejo el piso entonces a @FvD para que dé su opinión.

¡Muchas gracias @pachamaltese por los cambios, y @melvidoni por facilitar el proceso de revisión!
