### 2021-04-22 #414 @pachamaltese
> Hola a todos,
>
> He vuelto a hacer una revisión del paquete y confirmo que @pachamaltese ha incorporado mis principales comentarios. Por lo tanto, considero que el paquete está listo para ser aprovado tada
>
> Dejo el piso entonces a @FvD para que dé su opinión.
>
> ¡Muchas gracias @pachamaltese por los cambios, y @melvidoni por facilitar el proceso de revisión!
>
> **Edit:** No lo mencioné, pero me parece excelente haber incluido las dos nuevas tablas con la codificación y los valores de las variables. Es un gran aporte que seguramente va a facilitar mucho el uso para los usuarios smile

gracias !! si, es mejor que ver un XML que incluso a mi me resulta impactante
