+++ README.md
@@ -19,31 +19,31 @@ status](https://www.r-pkg.org/badges/version/censo2017)](https://CRAN.R-project.
 
 # Acerca de
 
-Provee un acceso conveniente a más de 17 millones de registros de la
+Provee un acceso conveniente a mas de 17 millones de registros de la
 base de datos del Censo 2017. Los datos fueron importados desde el DVD
 oficial del INE usando el [Convertidor
 REDATAM](https://github.com/discontinuos/redatam-converter/) creado por
-Pablo De Grande y además se proporcionan los mapas que acompañan a estos
-datos. Estos mismos datos en DVD posteriormente quedaron disponibles en
-las [Bases de Datos del
+Pablo De Grande y ademas se proporcionan los mapas que acompanian a
+estos datos. Estos mismos datos en DVD posteriormente quedaron
+disponibles en las [Bases de Datos del
 INE](https://www.ine.cl/estadisticas/sociales/censos-de-poblacion-y-vivienda/poblacion-y-vivienda).
 
-Después de la primera llamada a `library(censo2017)` se le pedirá al
+Despues de la primera llamada a `library(censo2017)` se le pedira al
 usuario que descargue la base usando `censo_descargar_base()` y se puede
 modificar la ruta de descarga con la variable de entorno
 `CENSO_BBDD_DIR`. La variable de entorno se puede crear con
 `usethis::edit_r_environ()`.
 
-El sitio de la documentación es pacha.dev/censo2017 y las viñetas con
-ejemplos están disponibles en pacha.dev/censo2017/articles.
+El sitio de la documentacion es pacha.dev/censo2017 y las vinietas con
+ejemplos estan disponibles en pacha.dev/censo2017/articles.
 
-# Instalación
+# Instalacion
 
-Versión estable
+Version estable
 
     install.packages("censo2017")
 
-Versión de desarrollo
+Version de desarrollo
 
     # install.packages("remotes")
     remotes::install_github("pachamaltese/censo2017")
@@ -52,36 +52,36 @@ Versión de desarrollo
 
 Esta version de la base de datos del Censo 2017 presenta algunas
 diferencias respecto de la original que se obtiene en DVD y corresponde
-a una versión DuckDB derivada a partir de los Microdatos del Censo 2017
+a una version DuckDB derivada a partir de los Microdatos del Censo 2017
 en formato DVD.
 
 La modificacion sobre los archivos originales, que incluyen geometrias
 detalladas disponibles en [Cartografias
 Censo2017](https://github.com/pachamaltese/cartografias-censo2017),
-consistió en unir todos los archivos SHP regionales en una ínica tabla
+consistio en unir todos los archivos SHP regionales en una unica tabla
 por nivel (e.g en lugar de proveer `R01_mapa_comunas`, …,
-`R15_mapa_comunas` combiné las 15 regiones en una única tabla
+`R15_mapa_comunas` combine las 15 regiones en una unica tabla
 `mapa_comunas`).
 
 Los cambios concretos respecto de la base original son los siguientes:
 
 -   Nombres de columna en formato “tidy” (e.g. `comuna_ref_id` en lugar
     de `COMUNA_REF_ID`).
--   Agregué los nombres de las unidades geográficas (e.g. se incluye
+-   Agregue los nombres de las unidades geograficas (e.g. se incluye
     `nom_comuna` en la tabla `comunas` para facilitar los filtros).
--   Añadí la variable `geocodigo` a la tabla de `zonas`. Esto facilita
+-   Aniadi la variable `geocodigo` a la tabla de `zonas`. Esto facilita
     mucho las uniones con las tablas de mapas en SQL.
[…]
