### 2021-04-25 #414 @pachamaltese
nuevos cambios logrados:

- [x] Agregar "La documentación del paquete y ejemplos de uso se encuentran en https://pacha.dev/censo2017/." al msje de inicio
- [x] Tildes en la documentación *escribi todo sin acentos, estilo los ninios y la ninias*
- [x] README: link a la viñeta
- [x] Explicar de mejor manera la interacción con chilemapas
- [x] Links a los multiples sitios del INE
- [x] Quitar el mensaje de cómo citar
- [x] Cambiar el mensaje que le pide al usuario leer la documentación (más detalles abajo)
- [x] `censo_borrar_base()` no elimina el directorio especificado en `censo_path()`
- [x] El test correspondiente no verifica que el directorio y sus contenidos hayan sido eliminados del disco
- [x] Falta guia de contribución en el README y/o archivo CONTRIBUTING.
- [x] Puede ser también una buena idea explicar en el README que después de la primera llamada a library(censo2017) se le pedirá al usuario que descargue la base usando censo_descargar_base(), y que puede modificar la ruta de descarga con la variable de entorno CENSO_BBDD_DIR. *Listo y con referencia a usethis::*
- [x] Sacar referencias a los mapas, pues las cartografías las movi a otro repositorio
- [ ] Pulida a nivel de documentación
- [x] Top level documentation
- [x] Usar codemetar::
- [x] Mensaje "Error, no esta disponible esta tabla, estas seguro que instalaste los datos con censo_descargar_base()`."
- [x] Extender o los ejemplos o la descripción de algunas funciones

Estos creo que se lograron adecuadamente, pero me gustaría saber si hace falta darle un mayor nivel de detalle
- [ ] Explicar requisitos de instalación
- [ ] Escribir propósito del paquete (para que usuario, etc)
