### 2021-04-25 #414 @melvidoni
Excelente! @pachamaltese el paquete está aprobado entonces. Te dejo el comenario en Inglés así es asequible a los otros editores.

---

Approved! Thanks @pachamaltese  for submitting and @mpaulacaldas @FvD  for your reviews!

To-dos:
    * replace your package docs URL with `https://docs.ropensci.org/package_name`
- [ ] Fix any links in badges for CI and coverage to point to the ropensci URL. We no longer transfer Appveyor projects to ropensci Appveyor account so after transfer of your repo to rOpenSci's "ropensci" GitHub organization the badge should be `[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/pkgname?branch=master&svg=true)](https://ci.appveyor.com/project/individualaccount/pkgname)`. If Appveyor does not pick up new commits after transfer, you might need to delete and re-create the Appveyor project. (Repo transfers are smoother with GitHub Actions)
- [ ] Please check you updated the package version post-review version updated and that you documented all changes in NEWS.md
