+++ CODE_OF_CONDUCT.md
@@ -1,5 +1,0 @@
-# Codigo de conducta
-
-El repositorio del paquete Censo 2017 es un espacio abierto que tiene por objetivo la generacion de recursos para la comunidad chilena e hispanohablante en general a traves de una dinamica de trabajo colaborativa. Durante el proceso de revision y mejora, se espera que las personas puedan participar libremente, compartiendo ideas, haciendo sugerencias, proponiendo cambios, analizando puntos de vista, etc., en un ambiente de respeto y colaboracion. Asimismo, se espera que quienes participen hagan criticas justas, constructivas y propositivas, no juicios de valor.
-
-En todo momento se cautelara que este sea un ambiente libre de acoso y hostigamiento, independiente del sexo, identidad, genero, edad, orientacion sexual, discapacidad, apariencia fisica, tamaño corporal, raza, etnia, religion (o la falta de ella), ideologia, nacionalidad, variante lingüistica u opciones tecnologicas. No se tolerara el acoso ni el hostigamiento en ninguna de sus formas. Quienes incurran en este tipo de conductas seran marginados del proceso de trabajo de forma permanente. El lenguaje sexual y las imagenes de ese tipo no son apropiados en ningun espacio de este proyecto.
