### 2021-08-13 #414 @pachadotdev
hola @melvidoni, los badges aún dicen "en revisión" ¿cómo se cambia eso?
