### 2021-08-15 #414 @melvidoni
Hola. El 26 de Abril lo marqué como "approved" a este issue. ¿A qué badge te referís?
