### 2021-09-30 #414 @pachadotdev
> Hola. El 26 de Abril lo marqué como "approved" a este issue. ¿A qué badge te referís?

mis disculpas por lo lento, no sé como lograr que github notifique estas cosas

acá sale como "under review"
![image](https://user-images.githubusercontent.com/10091065/135382505-4d8a0e11-267d-47e1-b5d4-7cd3fc71b367.png)
