### 2021-09-30 #414 @maelle
Por suerte en el futuro habrá menos razones de escribir el codigo del badge "a mano": si el editor o la editora añade les revisores con los comandos de bot https://devdevguide.netlify.app/editorguide.html#look-for-and-assign-reviewers, el bot verrá cuando ya hay 2 revisores, cambia el label y escribe un [comentario con el codigo para el badge](https://github.com/ropensci/software-review/blob/master/.buffy/templates/badge.md). :sweat_smile:
